Refactor matchers out into their own file for easier maintenance
