+++ test/support/html_matchers.ex
@@ -1,0 +1,50 @@
+defmodule AtomStyleTweaks.HtmlMatchers do
+  import ExUnit.Assertions
+
+  def find_element(conn, selector) do
+    conn
+    |> decoded_response(200)
+    |> Floki.find(selector)
+  end
+
+  def has_attribute(element, expected_attribute) do
+    assert get_attribute(element, expected_attribute)
+
+    element
+  end
+
+  def has_attribute(element, attribute, expected_value) do
+    assert get_attribute(element, attribute) == expected_value
+
+    element
+  end
+
+  def has_text(element, expected) do
+    assert get_text(element) == expected
+
+    element
+  end
+
+  def links_to(element, expected) do
+    assert get_href(element) == expected
+
+    element
+  end
+
+  def matches_text(element, expected) do
+    assert get_text(element) =~ expected
+
+    element
+  end
+
+  defp decoded_response(conn, status_code) do
+    conn
+    |> Phoenix.ConnTest.html_response(status_code)
+    |> HtmlEntities.decode
+  end
+
+  defp get_attribute(element, attribute) when is_atom(attribute), do: get_attribute(element, Atom.to_string(attribute))
+  defp get_attribute(element, attribute), do: hd(Floki.attribute(element, attribute))
+  defp get_text(element), do: Floki.text(element)
+  defp get_href(element), do: get_attribute(element, :href)
+end
